requirements.txt
fastapi
sentence-transformers
faiss-cpu
uvicorn